FullNetworkSystems

Documentation for FullNetworkSystems.

FullNetworkSystems.Branch

FullNetworkSystems.Bus

FullNetworkSystems.Generator

FullNetworkSystems.System

FullNetworkSystems.SystemDA

FullNetworkSystems.SystemRT

FullNetworkSystems.Zone

FullNetworkSystems.branches_by_breakpoints

FullNetworkSystems.gens_per_zone

FullNetworkSystems.get_datetimes

FullNetworkSystems.Branch — Type

struct Branch

Type for static branch attributes. Branches may have between 0 and 2 break points which is why the break_points and penalties fields contain variable length Tuples.

Fields:

name::String31

Branch long name

to_bus::String15

Name of the bus the branch goes to

from_bus::String15

Name of the bus the branch goes from

rate_a::Float64

Power flow limit for the base case (MVA)

rate_b::Float64

Power flow limit for contingency scenario (MVA)

is_monitored::Bool

Boolean defining whether the branch is monitored

break_points::Tuple{Float64, Float64}

Break points of the branch. Branches can have 0, 1, or 2 break points. Zeros indicate no break point

penalties::Tuple{Float64, Float64}

Price penalties for each of the break points of the branch ($)

FullNetworkSystems.Bus — Type

struct Bus

Type for static bus attributes.

Fields:

name::String15

Bus name

base_voltage::Float64

Base voltage (kV)

FullNetworkSystems.Generator — Type

struct Generator

Type for static generator attribute (i.e. things that describe a generator that are not time series data).

Fields:

unit_code::Int64

Generator id/unit code

zone::Int64

Number of the zone the generator is located in

startup_cost::Float64

Cost of turning on the generator ($)

shutdown_cost::Float64

Cost of turning off the generator ($)

no_load_cost::Float64

Cost of the generator being on but not producing any MW ($ /hour)

min_uptime::Float64

Minimum time the generator has to be committed for (hours)

min_downtime::Float64

Minimum time the generator has to be off for (hours)

ramp_up::Float64

Rate at which the generator can increase generation (MW/minute)

ramp_down::Float64

Rate at which the generator can decrease generation (MW/minute)

technology::Symbol

Symbol describing the technology of the generator

FullNetworkSystems.System — Type

System

The abstract type for representing the whole power system including topology, static components and their attributes, and time series data.

Topology: Dictionaries linking generators, loads, and bids (if present) to buses. System wide static components and grid matrices: zones, buses, generators, branches, LODF and PTDF. Time series data: all the time series associated with generators, loads and bids. All stored as KeyedArrays of ids x datetimes.

FullNetworkSystems.SystemDA — Type

struct SystemDA <: System

Subtype of a System for modelling the day-ahead market.

Fields:

gens_per_bus::Dictionaries.Dictionary{String15, Vector{Int64}}

Dictionary where the keys are bus names and the values are generator ids at that bus

incs_per_bus::Dictionaries.Dictionary{String15, Vector{String}}

Dictionary where the keys are bus names and the values are increment bid ids at that bus

decs_per_bus::Dictionaries.Dictionary{String15, Vector{String}}

Dictionary where the keys are bus names and the values are decrement bid ids at that bus

psds_per_bus::Dictionaries.Dictionary{String15, Vector{String}}

Dictionary where the keys are bus names and the values are price sensitive demand bid ids at that bus

loads_per_bus::Dictionaries.Dictionary{String15, Vector{String}}

Dictionary where the keys are bus names and the values are load ids at that bus

zones::Dictionaries.Dictionary{Int64, Zone}

Zones in the System, which will also include a Zone entry for the market wide zone

buses::Dictionaries.Dictionary{String15, Bus}

Buses in the System indexed by bus name

generators::Dictionaries.Dictionary{Int64, Generator}

Generators in the System indexed by unit code

branches::Dictionaries.Dictionary{String31, Branch}

Branches in the System indexed by branch name

LODF::Dictionaries.Dictionary{String, KeyedArray{Float64, 2}}

The line outage distribution factor matrix of the system for a set of contingencies given by the keys of the Dictionary. Each entry is a KeyedArray with axis keys branch names x branch on outage

PTDF::KeyedArray{Float64, 2}

Power transfer distribution factor of the system. KeyedArray where the axis keys are branch names x bus names

initial_generation::KeyedArray{Float64, 1}

Generation of the generators at the start of the day (MW)

hours_at_status::KeyedArray{Float64, 1}

Hours each generator has been at its current status at the start of the day

offer_curve::KeyedArray{Vector{Tuple{Float64, Float64}}, 2}

Generator offer curves. KeyedArray where the axis keys are generator names x datetimes

availability::KeyedArray{Bool, 2}

Generator availability

must_run::KeyedArray{Bool, 2}

Generator must run flag indicating that the generator has to be committed at that hour

regulation_min::KeyedArray{Float64, 2}

Generator minimum output in the ancillary services market (MW)

regulation_max::KeyedArray{Float64, 2}

Generator maximum output in the ancillary services market (MW)

pmin::KeyedArray{Float64, 2}

Generator minimum output (MW)

pmax::KeyedArray{Float64, 2}

Generator maximum output (MW)

asm_regulation::KeyedArray{Float64, 2}

Ancillary services regulation offer prices ($ /MW)

asm_spin::KeyedArray{Float64, 2}

Ancillary services spinning offer prices ($ /MW)

asm_sup_on::KeyedArray{Float64, 2}

Ancillary services supplemental on offer prices ($ /MW)

asm_sup_off::KeyedArray{Float64, 2}

Ancillary services supplemental off offer prices ($ /MW)

loads::KeyedArray{Float64, 2}

Load time series data. KeyedArray where the axis keys are load ids x datetimes

increment::KeyedArray{Vector{Tuple{Float64, Float64}}, 2}

Increment bids time series data. KeyedArray where the axis keys are bid ids x datetimes

decrement::KeyedArray{Vector{Tuple{Float64, Float64}}, 2}

Decrement bids time series data. KeyedArray where the axis keys are bid ids x datetimes

price_sensitive_demand::KeyedArray{Vector{Tuple{Float64, Float64}}, 2}

Price sensitive demand bids time series data. KeyedArray where the axis keys are bid ids x datetimes

FullNetworkSystems.SystemRT — Type

struct SystemRT <: System

Subtype of a System for modelling the real-time market.

Fields:

gens_per_bus::Dictionaries.Dictionary{String15, Vector{Int64}}

Dictionary where the keys are bus names and the values are generator ids at that bus

loads_per_bus::Dictionaries.Dictionary{String15, Vector{String}}

Dictionary where the keys are bus names and the values are load ids at that bus

zones::Dictionaries.Dictionary{Int64, Zone}

Zones in the System, which will also include a Zone entry for the market wide zone

buses::Dictionaries.Dictionary{String15, Bus}

Buses in the System indexed by bus name

generators::Dictionaries.Dictionary{Int64, Generator}

Generators in the System indexed by unit code

branches::Dictionaries.Dictionary{String31, Branch}

Branches in the System indexed by branch name

LODF::Dictionaries.Dictionary{String, KeyedArray{Float64, 2}}

The line outage distribution factor matrix of the system for a set of contingencies given by the keys of the Dictionary. Each entry is a KeyedArray with axis keys branch names x branch on outage

PTDF::KeyedArray{Float64, 2}

Power transfer distribution factor of the system. KeyedArray where the axis keys are branch names x bus names

initial_generation::KeyedArray{Float64, 1}

Generation of the generator at the start of the time period (MW)

offer_curve::KeyedArray{Vector{Tuple{Float64, Float64}}, 2}

Generator offer curves. KeyedArray where the axis keys are generator names x datetimes

status::KeyedArray{Bool, 2}

Generator status indicated by a Bool

status_regulation::KeyedArray{Bool, 2}

Generator regulation status indicated by a Bool

regulation_min::KeyedArray{Float64, 2}

Generator minimum output in the ancillary services market (MW)

regulation_max::KeyedArray{Float64, 2}

Generator maximum output in the ancillary services market (MW)

pmin::KeyedArray{Float64, 2}

Generator minimum output (MW)

pmax::KeyedArray{Float64, 2}

Generator maximum output (MW)

asm_regulation::KeyedArray{Float64, 2}

Ancillary services regulation offer prices ($ /MW)

asm_spin::KeyedArray{Float64, 2}

Ancillary services spinning offer prices ($ /MW)

asm_sup_on::KeyedArray{Float64, 2}

Ancillary services supplemental on offer prices ($ /MW)

asm_sup_off::KeyedArray{Float64, 2}

Ancillary services supplemental off offer prices ($ /MW)

loads::KeyedArray{Float64, 2}

Load time series data. KeyedArray where the axis keys are load ids x datetimes

FullNetworkSystems.Zone — Type

struct Zone

Type defining a market zone. The Zone is identified by a number. The other fields contain the service requirements for the zone.

Fields:

number::Int64

Zone number

reg::Float64

Zonal regulation requirement (MW)

spin::Float64

Zonal spinning requirement (MW)

sup_on::Float64

Zonal supplemental on requirement (MW)

sup_off::Float64

Zonal supplemental off requirement (MW)

FullNetworkSystems.branches_by_breakpoints — Method

branches_by_breakpoints(system::System)

Returns three vectors containing of the names of branches which have 0, 1, and 2 breakpoints.

FullNetworkSystems.gens_per_zone — Method

gens_per_zone(system::System)

Returns a Dict with keys of Zone numbers and values of generator names in that zone.

FullNetworkSystems.get_datetimes — Method

get_datetimes(system)

Extract datetimes from a System.